@@ -2,6 +2,7 @@
   "name": "app",
   "private": true,
   "version": "0.0.0",
+  "type": "module",
   "author": "trevorism <tbrooks@trevorism.com>",
   "scripts": {
     "build": "vite build",
@@ -9,32 +10,33 @@
     "accept": "vitest run",
     "dev": "vite",
     "preview": "vite preview",
+    "format": "prettier --write src/",
     "test": "vitest"
   },
   "dependencies": {
     "@vuestic/tailwind": "^0.1.3",
-    "vue": "^3.3.9",
+    "vue": "^3.3.13",
     "vue-router": "^4.2.5",
     "vue3-click-away": "^1.2.4",
-    "vuestic-ui": "^1.8.5"
+    "vuestic-ui": "^1.8.7"
   },
   "devDependencies": {
-    "@rushstack/eslint-patch": "^1.6.0",
-    "@trevorism/ui-header-bar": "^3.4.0",
-    "@vitejs/plugin-vue": "^4.5.1",
+    "@rushstack/eslint-patch": "^1.6.1",
+    "@trevorism/ui-header-bar": "^3.5.0",
+    "@vitejs/plugin-vue": "^4.5.2",
     "@vue/eslint-config-prettier": "^8.0.0",
     "@vue/test-utils": "^2.4.3",
     "autoprefixer": "^10.4.16",
     "axios": "^1.6.2",
-    "eslint": "^8.55.0",
+    "eslint": "^8.56.0",
     "eslint-plugin-vue": "^9.19.2",
     "jsdom": "^23.0.1",
     "material-design-icons-iconfont": "^6.7.0",
     "postcss": "^8.4.32",
-    "prettier": "^3.1.0",
-    "tailwindcss": "^3.3.5",
-    "vite": "^5.0.4",
-    "vitest": "^0.34.6",
+    "prettier": "^3.1.1",
+    "tailwindcss": "^3.4.0",
+    "vite": "^5.0.10",
+    "vitest": "^1.1.0",
     "vue-cli-plugin-vuestic-ui": "~1.0.11",
     "vue3-cookies": "^1.0.6"
   }
